:lipstick: (pc-17356) make Clock icon Bicolor
