+++ src/features/bookings/components/OnGoingBookingItem.tsx
@@ -8,7 +8,7 @@ import { useCategoryId, useSubcategory } from 'libs/subcategories'
 import { tileAccessibilityLabel, TileContentType } from 'libs/tileAccessibilityLabel'
 import { OfferImage } from 'ui/components/tiles/OfferImage'
 import { TouchableLink } from 'ui/components/touchableLink/TouchableLink'
-import { Clock as DefaultClock } from 'ui/svg/icons/Clock'
+import { BicolorClock as DefaultClock } from 'ui/svg/icons/BicolorClock'
 import { Duo } from 'ui/svg/icons/Duo'
 import { OfferEvent as DefaultOfferEvent } from 'ui/svg/icons/OfferEvent'
 import { getSpacing, Spacer, Typo } from 'ui/theme'
