+++ src/features/profile/components/Badges/IdentityCheckPendingBadge.tsx
@@ -2,11 +2,11 @@ import { t } from '@lingui/macro'
 import React from 'react'
 
 import { ProfileBadge } from 'features/profile/components/Badges/ProfileBadge'
-import { Clock } from 'ui/svg/icons/Clock'
+import { BicolorClock } from 'ui/svg/icons/BicolorClock'
 
 export const IdentityCheckPendingBadge = () => (
   <ProfileBadge
-    popOverIcon={Clock}
+    popOverIcon={BicolorClock}
     message={t`Ton inscription est en cours de traitement.`}
     testID={'identity-check-pending-badge'}
   />
