+++ src/features/profile/components/Badges/ProfileBadge.native.test.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { ProfileBadge } from 'features/profile/components/Badges/ProfileBadge'
 import { shouldOpenInbox } from 'features/profile/utils'
 import { render, fireEvent } from 'tests/utils'
-import { Clock } from 'ui/svg/icons/Clock'
+import { BicolorClock } from 'ui/svg/icons/BicolorClock'
 import { EmailFilled } from 'ui/svg/icons/EmailFilled'
 
 jest.mock('features/profile/utils')
@@ -16,7 +16,7 @@ describe('ProfileBadge', () => {
 
   it('should render component correctly if icon is provided', () => {
     const myComponent = render(
-      <ProfileBadge message={'Tu as déposé ton dossier. Bravo'} popOverIcon={Clock} />
+      <ProfileBadge message={'Tu as déposé ton dossier. Bravo'} popOverIcon={BicolorClock} />
     )
     expect(myComponent).toMatchSnapshot()
   })
@@ -25,7 +25,7 @@ describe('ProfileBadge', () => {
     const myComponent = render(
       <ProfileBadge
         message={'Tu as déposé ton dossier. Bravo'}
-        popOverIcon={Clock}
+        popOverIcon={BicolorClock}
         callToActionIcon={EmailFilled}
       />
     )
@@ -36,7 +36,7 @@ describe('ProfileBadge', () => {
     const myComponent = render(
       <ProfileBadge
         message={'Tu as déposé ton dossier. Bravo'}
-        popOverIcon={Clock}
+        popOverIcon={BicolorClock}
         callToActionIcon={EmailFilled}
         callToActionMessage={'Tu peux cliquer ici'}
         callToActionLink={'https://calltoaction.com'}
@@ -49,7 +49,7 @@ describe('ProfileBadge', () => {
     const myComponent = render(
       <ProfileBadge
         message={'Tu as déposé ton dossier. Bravo'}
-        popOverIcon={Clock}
+        popOverIcon={BicolorClock}
         callToActionIcon={EmailFilled}
         callToActionMessage={'Tu peux cliquer ici'}
       />
@@ -61,7 +61,7 @@ describe('ProfileBadge', () => {
     const { getByTestId } = render(
       <ProfileBadge
         message={'Tu as déposé ton dossier. Bravo'}
-        popOverIcon={Clock}
+        popOverIcon={BicolorClock}
         callToActionIcon={EmailFilled}
         callToActionLink={'https://google.com'}
         callToActionMessage={'Tu peux cliquer ici'}
