+++ src/features/profile/components/Badges/YoungerBadge.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 
 import { ProfileBadge } from 'features/profile/components/Badges/ProfileBadge'
 import { formatToSlashedFrenchDate } from 'libs/dates'
-import { Clock } from 'ui/svg/icons/Clock'
+import { BicolorClock } from 'ui/svg/icons/BicolorClock'
 
 export function YoungerBadge(props: { eligibilityStartDatetime: Date }) {
   const information = t({
@@ -13,5 +13,5 @@ export function YoungerBadge(props: { eligibilityStartDatetime: Date }) {
       'Patience\u00a0! Reviens à partir du {date} pour continuer ton inscription et bénéficier du crédit pass Culture.',
   })
 
-  return <ProfileBadge popOverIcon={Clock} message={information} testID={'younger-badge'} />
+  return <ProfileBadge popOverIcon={BicolorClock} message={information} testID={'younger-badge'} />
 }
