+++ src/features/profile/pages/ChangeEmail/AlreadyChangedEmailDisclaimer.tsx
@@ -3,7 +3,7 @@ import React, { useEffect } from 'react'
 import styled from 'styled-components/native'
 
 import { analytics } from 'libs/firebase/analytics'
-import { Clock } from 'ui/svg/icons/Clock'
+import { BicolorClock } from 'ui/svg/icons/BicolorClock'
 import { getSpacingString, Spacer, Typo } from 'ui/theme'
 
 export const AlreadyChangedEmailDisclaimer = () => {
@@ -30,8 +30,9 @@ const Container = styled.View(({ theme }) => ({
   borderRadius: theme.borderRadius.radius,
 }))
 
-const StyledClock = styled(Clock).attrs(({ theme }) => ({
+const StyledClock = styled(BicolorClock).attrs(({ theme }) => ({
   color: theme.colors.greyDark,
+  color2: theme.colors.greyDark,
 }))``
 
 const BodyText = styled(Typo.Caption)({
